requirements.txt
absl-py==0.1.10
bleach==1.5.0
certifi==2018.1.18
chardet==3.0.4
decorator==4.2.1
Django==2.0.2
djangorestframework==3.7.7
future==0.16.0
gevent==1.3a1
greenlet==0.4.13
html5lib==0.9999999
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.1
Markdown==2.6.11
msgpack-python==0.5.4
numpy==1.14.0
opencv-python==3.4.0.12
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
pkg-resources==0.0.0
prompt-toolkit==1.0.15
protobuf==3.5.1
ptyprocess==0.5.2
Pygments==2.2.0
pytz==2018.3
pyzmq==17.0.0
requests==2.18.4
simplegeneric==0.8.1
six==1.11.0
tensorflow==1.5.0
tensorflow-tensorboard==1.5.1
traitlets==4.3.2
urllib3==1.22
Wand==0.4.4
wcwidth==0.1.7
Werkzeug==0.14.1
zerorpc==0.6.1
